+++ src/rules/keyframe-declaration-no-important/__tests__/index.js
@@ -16,6 +16,10 @@ testRule(rule, {
   }, {
     code: "@keyframes important { from { margin: 1px; } }",
   }, {
+    code: "@-webkit-keyframes important { from { margin: 1px; } }",
+  },  {
+    code: "@-WEBKIT-KEYFRAMES important { from { margin: 1px; } }",
+  }, {
     code: "@non-keyframes important { from { margin: 1px !important; } }",
   }, {
     code: "@keyframes-non important { from { margin: 1px !important; } }",
@@ -27,6 +31,18 @@ testRule(rule, {
     message: messages.rejected,
     line: 1,
     column: 44,
+  }, {
+    code: "@-webkit-keyframes important { from { margin: 1px !important; } }",
+    description: "with !important",
+    message: messages.rejected,
+    line: 1,
+    column: 52,
+  }, {
+    code: "@-WEBKIT-KEYFRAMES important { from { margin: 1px !important; } }",
+    description: "with !important",
+    message: messages.rejected,
+    line: 1,
+    column: 52,
   }, {
     code: "@keyframes important { from { margin: 1px!important; } }",
     description: "with !important",